Still work in progress.
I had to move all the Vix code to its own file and structure so the code
looks cleaner and easier to maintain. There was a lot of noise
going on inside Terraform resource functions due to data mapping
and plumbing required to get the configuration attributes ready to use
by Vix.