Implements resource update function and adds more fixes
It also makes sure to set the vm name and creates a new private function
to power off virtual machines in the best way possible.

The vix_to_tf data mapping function was comented as refreshing the resource
should not undefined attributes, also it is not possible to identify
whether a boolean value changed or not. So, refreshing a resource should
only set attributes that makes sense in the vm runtime context. This mean
that attributes whose purpose is to modify the behavior of the provider
should not be attempted to refresh at all or the resource will fall
under an inconsistent state.