Forces VM removal ignoring existing locks it might have.
When launching a VM with GUI enabled, the destroy action fails due to
the GUI holding a lock onto the VM. This change forces the removal of
any locks the VM might have and deletes the VM. A side effect caused by this
change is the VMware UI being in an inconsistent state where you would still
see the VM listed but unable to open it. Just go head and remove it
manually from list manually and everything will be alright.