Creates linked clones instead of unpacking the image multiple times.
This fixes an issue happening when launching multiple virtual machines
using the same vmx file. Instead, we now download the image and unpack
it into a folder for only Gold images. Gold images are never launched and only
exists so that we can clone new virtual machines from it. If a gold virtual
machine is deleted, all the linked clones that were created from it will be
invalid. The benefit of using linked clones as opposed to full clones
are listed here:
https://www.vmware.com/support/ws55/doc/ws_clone_overview.html