Fixes checking for an existing cloned.
It also fixes glob pattern so that it finds a vmx file not within a
nested directoy, in addition to finding it if it does exists within
nested folders. The reason we are doing this is that image archives
could sometimes come with nested folders and sometimes they wouldn't
have a folder but the mere files.